invoice-matcher/requirements.txt
PyMuPDF
pdfplumber
Pillow
google-generativeai
python-dotenv
streamlit
fpdf